requirements.txt

flask==3.0.2
playwright==1.43.0
python-dotenv==1.0.0
aiohttp==3.9.5
pytest==8.2.0
pytest-asyncio==0.23.7
ruff==0.5.7
mypy==1.10.0
requests==2.31.0
